Non-integer shapes will be unsupported in a future version of Numpy.

+++ astropy/convolution/convolve.py
@@ -435,7 +435,7 @@ def convolve_fft(array, kernel, boundary='fill', fill_value=0, crop=True,
             # add the shape lists (max of a list of length 4) (smaller)
             # also makes the shapes square
             fsize = 2 ** np.ceil(np.log2(np.max(arrayshape + kernshape)))
-        newshape = np.array([fsize for ii in range(array.ndim)])
+        newshape = np.array([fsize for ii in range(array.ndim)], dtype=int)
     else:
         if psf_pad:
             # just add the biggest dimensions
